Created Influence Map
Added an extra menu listing on the menubar with a boolean that if selected
draws an influence map on the board. Spaces turn blue if under strong black
influence, yellow if under strong white influence.

This is very naive, as a strong piece is actually one that's alive, regardless
of its surrounded. A better method would be to test life/death of a luster,
then weight spaces near live clusters as even more strongly influenced than
spaces next to undecided clusters. But for teaching influence of pieces on the
board, this implementation works pretty well.