Made it easier for cpu to pass
CPU can pass by returning two -1's. But CPU will only pass if there are no possible plays, and can not pass if black has already passed. So, modify the pass so that the actual game (through a menu setting) knows if there's a computer player. Then pass can call play() if there's a computer player, and see if white decides its time to pass.

Also, we need a much stronger way to determine if it's time for white to pass. This should probably be through a canMakeLife(), which looks at a group and determines if there's a way to create life for that group.